AI/Old/requirements.txt
numpy
httpx
requests
langchain
langdetect
transformers
python-dotenv
langchain-core
huggingface-hub
langchain_classic
langchain-huggingface
langchain_community
tqdm
pydantic
orjson
PyYAML
tenacity
regex
trafilatura
lxml_html_clean
opencv-python
pywavelets
scipy
pillow
piexif
huggingface_hub
basicsr
facexlib
realesrgan
torch
dotenv 
langchain_huggingface
langchain_core
mediapipe
YOLO
torchvision
timm
scikit-image
scikit-learn
joblib
lightgbm
exifread
pandas
torchaudio